# docs/swagger.yaml
basePath: /
definitions:
...
  models.Bahan:
    properties:
      batas_minimum:
        type: number
      harga_saat_ini:
        description: Update otomatis dari pembelian terakhir
        type: number
      id:
        type: integer
      nama_bahan:
        type: string
      satuan:
        description: gr, ml, pcs
        type: string
      stok:
        type: number
      urutan:
        type: integer
    type: object
  models.BahanInput:
    properties:
      batas_minimum:
        example: 5000
        type: number
      harga_saat_ini:
        example: 12500
        type: number
      nama_bahan:
        example: Tepung Cakra Kembar
        type: string
      satuan:
        example: gr
        type: string
      stok:
        example: 15000
        type: number
    required:
    - nama_bahan
    - satuan
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: Terjadi kesalahan pada server
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: Operasi berhasil dilakukan
        type: string
    type: object
...
  models.Resep:
    properties:
      bahan_detail:
        items:
          $ref: '#/definitions/models.ResepBahan'
        type: array
      id:
        type: integer
      nama_resep:
        type: string
      target_gramasi:
        description: Total adonan matang dr 1 Resep
        type: number
    type: object
  models.ResepBahan:
    properties:
      bahan:
        $ref: '#/definitions/models.Bahan'
      bahan_id:
        type: integer
      id:
        type: integer
      kebutuhan:
        description: Butuh berapa gr/ml/pcs
        type: number
      resep_id:
        type: integer
    type: object
  models.ResepBahanInput:
    properties:
      bahan_id:
        example: 2
        type: integer
      kebutuhan:
        example: 250.5
        type: number
    required:
    - bahan_id
    - kebutuhan
    type: object
  models.ResepInput:
    properties:
      bahan_detail:
        items:
          $ref: '#/definitions/models.ResepBahanInput'
        type: array
      nama_resep:
        example: Adonan Roti Manis
        type: string
      target_gramasi:
        example: 2000
        type: number
    required:
    - bahan_detail
    - nama_resep
    - target_gramasi
    type: object
...
  models.UrutanBahanInput:
    properties:
      id:
        example: 1
        type: integer
      urutan:
        example: 2
        type: integer
    required:
    - id
    - urutan
    type: object
...
host: localhost:3000
info:
  contact:
    name: Aristo Budiman
  description: Dokumentasi API Internal untuk Sistem Inventory, Nota, dan Kas.
  title: Tiara Bakery Master API
  version: "1.0"
paths:
...
  /api/bahan:
    get:
      consumes:
      - application/json
      description: Mengambil daftar master bahan baku gudang, diurutkan berdasarkan
        urutan tampilan.
      produces:
      - application/json
      responses:
        "200":
          description: List data bahan berhasil ditarik
          schema:
            items:
              $ref: '#/definitions/models.Bahan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil Seluruh Master Bahan
      tags:
      - 05. Master Bahan Baku
    post:
      consumes:
      - application/json
      description: Menyimpan data master bahan baku baru ke gudang.
      parameters:
      - description: Data Bahan utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BahanInput'
      produces:
      - application/json
      responses:
        "200":
          description: Bahan berhasil dibuat
          schema:
            $ref: '#/definitions/models.Bahan'
        "400":
          description: Format JSON tidak valid
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Kesalahan eksekusi database
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat Master Bahan Baru
      tags:
      - 05. Master Bahan Baku
  /api/bahan/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus master bahan baku dari daftar aktif.
      parameters:
      - description: ID Bahan
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Bahan berhasil dihapus
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus Master Bahan (Soft Delete)
      tags:
      - 05. Master Bahan Baku
    put:
      consumes:
      - application/json
      description: Memperbarui profil, satuan, atau harga bahan baku berdasarkan ID.
      parameters:
      - description: ID Bahan
        in: path
        name: id
        required: true
        type: integer
      - description: Data update bahan utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.BahanInput'
      produces:
      - application/json
      responses:
        "200":
          description: Bahan berhasil diupdate
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Bahan tidak ditemukan
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Master Bahan
      tags:
      - 05. Master Bahan Baku
  /api/bahan/reorder:
    put:
      consumes:
      - application/json
      description: Memperbarui posisi urutan bahan baku untuk tampilan di form.
      parameters:
      - description: 'Array of object key: id, urutan'
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/models.UrutanBahanInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Urutan bahan berhasil diperbarui
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update Urutan Tampilan Bahan
      tags:
      - 05. Master Bahan Baku
...
  /api/resep:
    get:
      consumes:
      - application/json
      description: Mengambil daftar master resep beserta detail array komponen bahan-bahannya.
      produces:
      - application/json
      responses:
        "200":
          description: List data resep ditarik
          schema:
            items:
              $ref: '#/definitions/models.Resep'
            type: array
      security:
      - BearerAuth: []
      summary: Ambil Seluruh Master Resep
      tags:
      - 06. Master Resep
    post:
      consumes:
      - application/json
      description: Menyimpan data master resep baru lengkap dengan komposisi bahannya.
      parameters:
      - description: Data Resep utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ResepInput'
      produces:
      - application/json
      responses:
        "200":
          description: Resep berhasil dibuat
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Format JSON salah
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat Master Resep Baru
      tags:
      - 06. Master Resep
  /api/resep/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus master resep dari daftar aktif.
      parameters:
      - description: ID Resep
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Resep berhasil dihapus
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Hapus Master Resep (Soft Delete)
      tags:
      - 06. Master Resep
    put:
      consumes:
      - application/json
      description: Memperbarui data resep dan list komposisi bahan berdasarkan ID
        resep.
      parameters:
      - description: ID Resep
        in: path
        name: id
        required: true
        type: integer
      - description: Data Resep utuh
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.ResepInput'
      produces:
      - application/json
      responses:
        "200":
          description: Resep berhasil diupdate
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update Master Resep
      tags:
      - 06. Master Resep
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: "Bearer {token_jwt}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"